Hex widget integration.
